requirements.txt
streamlitstreamlit
pillow
pytesseract